% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmissionIntensities.R
\name{calcEmissionIntensities}
\alias{calcEmissionIntensities}
\title{Calculate Emission Intensities}
\usage{
calcEmissionIntensities(source_folder = NULL)
}
\arguments{
\item{source_folder}{Folder where FAOSTAT data (unzipped) is saved}
}
\value{
Emission intensities for IMPACT activities
}
\description{
Calculate Emission Intensities
}
\examples{
\dontrun{
calcEmissionIntensities(source_folder)
}
}
\author{
Abhijeet Mishra
}
