% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_fao.R
\name{clean_fao}
\alias{clean_fao}
\title{Cleanup fao data}
\usage{
clean_fao(f)
}
\arguments{
\item{f}{Name of file downloaded from FAOSTAT}
}
\value{
Cleaned FAO data
}
\description{
Cleanup fao data
}
\examples{
\dontrun{
clean_fao(f = "Emissions_crops_E_All_Data_(Normalized).csv")
}
}
\author{
Abhijeet Mishra
}
