% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{mapping}
\alias{mapping}
\title{Function to return mapping for IMPACT results}
\usage{
mapping(type = "j")
}
\arguments{
\item{type}{which mapping to return (j or c or cty to region)}
}
\value{
Mapping set (activity, commodity etc,)
}
\description{
Function to return mapping for IMPACT results
}
\examples{
\dontrun{
mapping(type = "j")
}
}
\author{
Abhijeet Mishra
}
