% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEmissions.R
\name{readEmissions}
\alias{readEmissions}
\title{Read Emission data from FAO}
\usage{
readEmissions(indicator = "crops", source_folder = NULL)
}
\arguments{
\item{indicator}{Options are "crop" emissions (reads file
"Emissions_crops_E_All_Data_(Normalized).csv"), "livestock" emissions (reads
file "Emissions_livestock_E_All_Data_(Normalized).csv")), and "total" emissions
(reads file "Emissions_Totals_E_All_Data_(Normalized).csv").
These files are all part of the bulk download from FAOSTAT.}

\item{source_folder}{Folder where FAOSTAT data (unzipped) is saved}
}
\value{
Emissions from FAO
}
\description{
Read Emission data from FAO
}
\examples{
\dontrun{
readEmissions(indicator = "crops")
}
}
\author{
Abhijeet Mishra
}
